Add skeleton for Watch()ing - move modules out, too
